requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
configparser==5.3.0
DBUtils==3.0.3
dnspython==2.3.0
exceptiongroup==1.1.3
Faker==18.13.0
Flask==2.2.5
gevent==22.10.2
greenlet==3.0.1
h11==0.14.0
idna==3.4
importlib-metadata==6.7.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpath==0.82.2
MarkupSafe==2.1.3
outcome==1.3.0.post0
packaging==23.2
pluggy==1.2.0
py==1.11.0
pycparser==2.21
pycryptodome==3.19.0
pymongo==4.6.0
PyMySQL==1.1.0
PySocks==1.7.1
pytest==7.4.3
pytest-html==3.2.0
pytest-metadata==3.0.0
python-dateutil==2.8.2
PyYAML==6.0.1
pyyaml-include==1.3.1
redis==5.0.1
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.15.0
zope.event==5.0
zope.interface==6.1